requirements.txt
flet>=0.21.0
cryptography>=42.0.0
Pillow>=10.0.0
pypdf>=4.0.0
tqdm>=4.66.0